MAINT: Move usecol handling to C and support more than integer cols
Of course to actually use that many columns you need A LOT of memory
right now.  Each field stores at least a UCS4 NUL character, but
the field is padded enough to require 16 bytes.
We always parse a full row, so that requires 20 bytes per field...
(i.e. 32 GiB RAM is not enough to test this :)).

+++ numpy/core/src/multiarray/textreading/rows.h
@@ -13,8 +13,8 @@
 
 NPY_NO_EXPORT PyArrayObject *
 read_rows(stream *s,
-        npy_intp nrows, int num_field_types, field_type *field_types,
-        parser_config *pconfig, int num_usecols, int *usecols,
+        npy_intp nrows, Py_ssize_t num_field_types, field_type *field_types,
+        parser_config *pconfig, Py_ssize_t num_usecols, Py_ssize_t *usecols,
         Py_ssize_t skiplines, PyObject *converters,
         PyArrayObject *data_array, PyArray_Descr *out_descr,
         bool homogeneous);
